word_embed_fineishes

+++ preprocessing/preprocessing.py
@@ -10,7 +10,11 @@ import preprocessFunctions as pfs
 import os
 import descriptiveFunctions as dfs
 from preprocessFunctions import add_vector_similarities
+import gensim.downloader as gensim_api
+from ast import literal_eval
 
+# Only needs to be executed once
+w2v_model = gensim_api.load('word2vec-google-news-300')
 
 ##################         ##################
 ##################functions##################
@@ -41,7 +45,6 @@ def main():
 
     pddf = pd.read_csv("resources/normalized_pd.csv", encoding="latin1")
     pfs.turnStringFieldToList(pddf, 'normalized_pd')
-    
 
 
     #normalize queries and product title
@@ -51,20 +54,21 @@ def main():
         normalizeCSV(qpdf, 'product_title')     
         qpdf.to_csv("resources/normalized_qp.csv") 
         
-    qpdf = pd.read_csv("resources/normalized_qp.csv", encoding="latin1")
-    pfs.turnStringFieldToList(qpdf, 'normalized_st')
-    pfs.turnStringFieldToList(qpdf, 'normalized_title')
+    # qpdf = pd.read_csv("resources/normalized_qp.csv", encoding="latin1")
+    # pfs.turnStringFieldToList(qpdf, 'normalized_st')
+    # pfs.turnStringFieldToList(qpdf, 'normalized_title')
     
 
     #make qfscore table
     if not os.path.exists("resources/qf_scores.csv"):
         qfScores = pfs.createQFScores(qpdf)
         qfScores.to_csv("resources/qf_scores.csv")
-    
+    #idfscore table
     if not os.path.exists("resources/qp_idf_scores.csv"):
         qp_idf_Scores = pfs.createIDFScores(qpdf.drop_duplicates(subset=['product_uid']), 'normalized_title')
         qp_idf_Scores.to_csv("resources/qp_idf_scores.csv")
-
+        
+    #for proximity score feature
     if not os.path.exists("resources/word_positions.csv"):
         print('making word_posts')
         pos_lists = pfs.createPosLists(pddf)
@@ -80,18 +84,46 @@ def main():
     #     pdIDFScores = pfs.createIDFScores(pddf, 'normalized_pd')
     #     pdIDFScores.to_csv("resources/pd_idf_scores.csv")
 
-        # Voeg samen:
-        merged = qpdf.merge(pddf[['product_uid', 'normalized_pd']], on='product_uid', how='left')
-        merged = merged.merge(qfScores, how='cross')
-        merged = merged.merge(qp_idf_Scores, how='cross')
-        merged = merged.merge(pd_idf_Scores, how='cross')
 
+    #word embedding
+    if not os.path.exists("resources/qp_with_vecsim.csv"):
+        # Specify the path to your CSV file
+        csv_path = "resources/query_product.csv"
+
+        # Read the CSV file into a pandas DataFrame
+        df = pd.read_csv(csv_path, encoding="latin1")
+        
+        nqpdf = pd.read_csv("resources/normalized_qp.csv")
+        npddf = pd.read_csv("resources/normalized_pd.csv")
+        df = df.join(nqpdf.set_index('id'), on='id')
+        df = df.join(npddf.set_index('product_uid'), on='product_uid')
+        df['normalized_st'] = df['normalized_st'].apply(literal_eval)
+        df['normalized_title'] = df['normalized_title'].apply(literal_eval)
+        df['normalized_pd'] = df['normalized_pd'].apply(literal_eval)
+
+        # Voeg samen:
+        #ik kreeg hier een merge error maar weet niet waarom
+        #dus ik heb ff mijn manier van mergen gedaan bovenaan in main
+        # merged = qpdf.merge(pddf[['product_uid', 'normalized_pd']], on='product_uid', how='left')
+        merged = df
+
+        #alles mergen in één grote tabel hoeft pas in main
+        # merged = merged.merge(qfScores, how='cross')
+        # merged = merged.merge(qp_idf_Scores, how='cross')
+        # merged = merged.merge(pd_idf_Scores, how='cross')
+        
         # Voeg vector features toe
         merged = add_vector_similarities(merged, w2v_model)
-
+        
+        #verwijder overbodige kolommen
+        merged = merged.drop('product_uid', axis=1)
+        merged = merged.drop('search_term', axis=1)
+        merged = merged.drop('relevance', axis=1)
+        merged = merged.drop('normalized_title', axis=1)
+        merged = merged.drop('normalized_st', axis=1)
+        merged = merged.drop('normalized_pd', axis=1)
         # Bewaar alles:
         merged.to_csv("resources/qp_with_vecsim.csv", index=False)
-
 
 main()
